% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{close_connection}
\alias{close_connection}
\title{close_connection}
\usage{
close_connection(connection)
}
\arguments{
\item{connection}{a connection created by
\code{\link[=create_connection]{create_connection()}}.}
}
\description{
Disconnect from the server.
}
\details{
This function closes the connection to the BOINC server and frees up busy
resources. In the case of ssh, it just calls the
\code{\link[ssh:ssh_disconnect]{ssh_disconnect()}} function from the ssh
package. In the case of the http/https interface, it looks for a link to the
exit page on the home page and follows it.
\subsection{Errors and warnings}{

When errors occur, execution can be stopped with the following messages:
\itemize{
\item for http/https connections:
\itemize{
\item "Already disconnected."
}
}
}
}
\examples{
\dontrun{

# For ssh connection:
con = create_connection(server = "ssh://boinc.local",
                        dir = "~/projects/myproject",
                        username = "boincadm",
                        password = "0000")

# For http connections:
con = create_connection(server = "http://boinc.local",
                        dir = "myproject",
                        username = "submitter@example.com",
                        password = "000000")

# For https connections:
con = create_connection(server = "https://boinc.local",
                        dir = "myproject",
                        username = "submitter@example.com",
                        password = "000000")
...

# Release resources:
close_connection(con)
}
}
