% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{create_connection}
\alias{create_connection}
\title{create_connection}
\usage{
create_connection(
  server,
  dir = "~/projects/rboinc",
  username,
  password = NULL,
  keyfile = NULL
)
}
\arguments{
\item{server}{a ssh or http server URL of the form
\code{<}protocol\code{>}://\code{<}server name\code{>}:\code{<}port\code{>}.
Examples:
\itemize{
\item "ssh://boinc-server.local" - for ssh connection;
\item "http://boinc-server.local" - for http connection;
\item "https://boinc-server.local" - for https connection.
}}

\item{dir}{a rboinc project directory on the server. For ssh connection,
this is the directory where the BOINC project is located. For http/https
connection, this is the full path to the project page (without the server
name).}

\item{username}{a string containing username. For ssh connection, this is the
user login. For http/https connection, this is the user email.}

\item{password}{a string containing user password. If this parameter is equal
to NULL, then a window will be displayed prompting you to enter the password}

\item{keyfile}{path to a private key file. For ssh connection only.}
}
\value{
A connection (list) for use by other functions.
}
\description{
Create a connection to the BOINC server.
}
\details{
This function create a connection to BOINC server. Don't edit the list
returned by this function.

Returned object contains valid information only until the end of the current
R session, so don't save it to a file. Before ending the session, it is
recommended to release the occupied resources by calling
\code{\link[=close_connection]{close_connection()}}.

\strong{ATTENTION:} ssh connections are supported but not recommended. If you
are the administrator for your BOINC project, see
https://boinc.berkeley.edu/trac/wiki/MultiUser for more information on the
http BOINC multiuser interface.
\subsection{Errors and warnings}{

When errors occur, execution can be stopped with the following messages:
\itemize{
\item for any connections:
\itemize{
\item "Unsupported server address format."
\item "Connection was canceled by user."
\item "Unrecognized protocol: "\code{<}protocol\code{>}""
}
\item for http/https connections:
\itemize{
\item "Authorization failed."
}
\item for ssh connections:
\itemize{
\item "Project directory was not found on server."
}
}
}
}
\examples{
\dontrun{

# For ssh connection:
con = create_connection(server = "ssh://boinc.local",
                        dir = "~/projects/myproject",
                        username = "boincadm",
                        password = "0000")

# For http connections:
con = create_connection(server = "http://boinc.local",
                        dir = "myproject",
                        username = "submitter@example.com",
                        password = "000000")

# For https connections:
con = create_connection(server = "https://boinc.local",
                        dir = "myproject",
                        username = "submitter@example.com",
                        password = "000000")
...

# Release resources:
close_connection(con)
}
}
