% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testAPI.R
\name{test_jobs}
\alias{test_jobs}
\title{test_jobs}
\usage{
test_jobs(
  work_func,
  data,
  init_func = NULL,
  global_vars = NULL,
  packages = c(),
  files = c(),
  callback_function = NULL
)
}
\arguments{
\item{work_func}{data processing function. This function runs for each element in data.}

\item{data}{data for processing.  Must be a numerable list or vector.}

\item{init_func}{initialization function.}

\item{global_vars}{a list in the format <variable name>=<value>.}

\item{packages}{a string vector with imported packages names.}

\item{files}{a string vector with the filenames that should be available for jobs.}

\item{callback_function}{a function that is called for each result after loading.}
}
\value{
a list with states of jobs. This list contains the following fields:
\itemize{
\item log - Rscript output;
\item result - computation result.
}
}
\description{
Like create_jobs, it creates a job for the BOINC server but does not submit it. Instead, it runs the job locally and generates a report at each step. This function is intended for debugging applications that use RBOINC. Files created by this function are not deleted after its completion.
}
