% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testAPI.R
\name{test_n_jobs}
\alias{test_n_jobs}
\title{test_n_jobs}
\usage{
test_n_jobs(
  work_func,
  data,
  n,
  init_func = NULL,
  global_vars = NULL,
  packages = c(),
  files = c(),
  callback_function = NULL
)
}
\arguments{
\item{work_func}{data processing function. This function runs for each element in data.}

\item{data}{data for processing.  Must be a numerable list or vector.}

\item{init_func}{initialization function.}

\item{global_vars}{a list in the format <variable name>=<value>.}

\item{packages}{a string vector with imported packages names.}

\item{files}{a string vector with the filenames that should be available for jobs.}

\item{callback_function}{a function that is called for each result after loading.}
}
\value{
a list with states of jobs. This list contains the following fields:
\itemize{
\item log - Rscript output;
\item result - computation result.
}
}
\description{
Like create_n_jobs, it creates a jobs for the BOINC server but does not submit them. Instead, it runs all jobs locally and generates a report at each step. This function is intended for debugging applications that use RBOINC. Files created by this function are not deleted after its completion.
}
\examples{
# import library
library(RBOINC.cl)
# function for processing data
fun = function(val)
{
  return(val * a + b)
}
# global variables
glob_vars = list(a = 3, b = 2)
# Initialization function. This function runs on each node for one times.
init = function()
{
  return(NULL)
}
# data for processing
data = list(matrix(rexp(15), 3,5), matrix(rexp(15), 3,5))

#callback function
print_func = function(val)
{
  print(val)
  # May be any value
  return(val)
  #return(NULL)
}

# Test jobs before sending
jobs_t = test_jobs(fun, data, init, glob_vars, c("httr", "xml2"), callback_function = print_func)
jobs_t

# Create connection:
#con = create_connection("ssh://boinc-server.local", dir = "~/projects/myproject", username = "boincadm", password = "0000") # ssh
#con = create_connection("http://boinc-server.local", dir = "myproject", username = "submitter@example.com", password = "000000")# http
con
# send jobs:
jobs = create_jobs(con, fun, data, init, glob_vars)
jobs
# Get jobs status. Run this until status not equal "done":
jobs = update_jobs_status(con, jobs)
jobs
# Close connection:
close_connection(con)
}
