% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancelJob.R
\name{cancel_jobs}
\alias{cancel_jobs}
\title{cancel_jobs}
\usage{
cancel_jobs(connection, jobs_status)
}
\arguments{
\item{connection}{a connection created by create_connection.}

\item{jobs_status}{a list returned by create_jobs or update_jobs_status.}
}
\value{
a list with current states of jobs. This list contains the following
fields:
\itemize{
\item jobs_name - a name of job on BOINC server;
\item results - computation results (NULL if computation is still incomplete).
The length of this list is equal to the length of the data;
\item jobs_status - jobs human-readable status for each job;
\item jobs_code - jobs status code, don't use this field;
\item status - computation status, may be:
\itemize{
\item "initialization" - jobs have been submitted to the server, but their
status was not requested by update_jobs_status.
\item "in_progress" - BOINC serves jobs.
\item "done" - computations are complete, the result was downloaded.
\item "error" - an error occurred during the job processing.
\item "queued" - job in the queue (only for http/https connections).
}
}
}
\description{
Cancel running jobs.
}
\details{
When errors occur, execution can be stopped with the following messages:
\itemize{
\item for unknown connections:
\itemize{
\item "Unknown protocol."
}
\item for any connection:
\itemize{
\item "All results have already been received."
}
}
}
