% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{close_connection}
\alias{close_connection}
\title{close_connection}
\usage{
close_connection(connection)
}
\arguments{
\item{connection}{a connection created by create_connection}
}
\description{
Disconnect connection.
}
\examples{
# import library
library(RBOINC.cl)
# function for processing data
fun = function(val)
{
  return(val * a + b)
}
# global variables
glob_vars = list(a = 3)
# initialize function
init = function()
{
  b <<- 2
}
# data for processing
data = list(matrix(rexp(15), 3,5), matrix(rexp(15), 3,5))

# Create connection:
#con = create_connection("ssh://boinc-server.local", dir = "~/projects/myproject", username = "boincadm", password = "0000") # ssh
#con = create_connection("http://boinc-server.local", dir = "myproject", username = "submitter@example.com", password = "000000")# http
con
# send jobs:
jobs = create_jobs(con, fun, data, init, glob_vars)
jobs
# Get jobs status. Run this until status not equal "complete":
jobs = update_jobs_status(con, jobs)
jobs
# Close connection:
close_connection(con)
}
