% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{close_connection}
\alias{close_connection}
\title{close_connection}
\usage{
close_connection(connection)
}
\arguments{
\item{connection}{a connection created by create_connection}
}
\description{
Disconnect from the server.
}
