% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{create_connection}
\alias{create_connection}
\title{create_connection}
\usage{
create_connection(
  host,
  dir = "~/projects/rboinc",
  password = NULL,
  keyfile = NULL
)
}
\arguments{
\item{host}{a ssh server string of the form <user@>hostname<:@port>.}

\item{dir}{a rboinc project dir on server.}

\item{password}{a string containing user password.}

\item{keyfile}{path to private key file.}
}
\value{
a connection for use by other functions
}
\description{
Create a ssh session.
}
