% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{create_connection}
\alias{create_connection}
\title{create_connection}
\usage{
create_connection(
  server,
  dir = "~/projects/rboinc",
  username,
  password = NULL,
  keyfile = NULL
)
}
\arguments{
\item{server}{an ssh or http server URL of the form <protocol>://<server name>:<port>. Examples:
\itemize{
\item "ssh://boinc-server.local" - for ssh connection;
\item "http://boinc-server.local" - for http connection;
\item "https://boinc-server.local" - for https connection.
}}

\item{dir}{a rboinc project directory on the server. For ssh connection, this is the directory where the BOINC project is located.
For http/https connection, this is the full path to the project page (without the server name).}

\item{username}{a string containing username. For ssh connection, this is the user login. For http/https connection, this is
the user email.}

\item{password}{a string containing user password.}

\item{keyfile}{path to a private key file. For ssh connection only.}
}
\value{
a connection (list) for use by other functions.

When errors occur, the following exceptions may be thrown:
\itemize{
\item for any connections:
\itemize{
\item Unsupported server address format.
\item Connection was canceled by user.
\item Unrecognized protocol: "<protocol>"
}
\item for http/https connections:
\itemize{
\item Authorization failed.
}
\item for ssh connections:
\itemize{
\item Project directory was not found on server.
\item Other exceptions thrown by ssh_connect.
}
}
}
\description{
Create an ssh or http/https connection to the BOINC server.
}
