% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{create_connection}
\alias{create_connection}
\title{create_connection}
\usage{
create_connection(
  host,
  dir = "~/projects/rboinc",
  password = NULL,
  keyfile = NULL
)
}
\arguments{
\item{host}{a ssh server string of the form <user@>hostname<:@port>.}

\item{dir}{a rboinc project dir on server.}

\item{password}{a string containing user password.}

\item{keyfile}{path to private key file.}
}
\value{
a connection (list) for use by other functions.
}
\description{
Create a ssh session.
}
\examples{
# import library
library(RBOINC.cl)
# function for processing data
fun = function(val)
{
  return(val * a + b)
}
# global variables
glob_vars = list(a = 3)
# initialize function
init = function()
{
  b <<- 2
}
# data for processing
data = list(matrix(rexp(15), 3,5), matrix(rexp(15), 3,5))

# Create connection:
con = create_connection("boincadm@boinc-server.local", dir = "~/projects/myproject", password = "0000")
con
# send jobs:
jobs = create_jobs(con, fun, data, init, glob_vars)
jobs
# Get jobs status. Run this until status not equal "complete":
jobs = update_jobs_status(con, jobs)
jobs
# Close connection:
close_connection(con)
}
