% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResult.R
\name{update_jobs_status}
\alias{update_jobs_status}
\title{update_jobs_status}
\usage{
update_jobs_status(connection, jobs_status, callback_function = NULL)
}
\arguments{
\item{connection}{a connection created by
\link[=create_connection]{create_connection}}

\item{jobs_status}{a list returned by
\link[=create_jobs]{create_jobs} or update_jobs_status.}

\item{callback_function}{a function with prototype
\code{function(result_element)} that is called for each result after loading.
The value returned by this function is placed in the result list.}
}
\value{
a list with current states of jobs. This list contains the following
fields:
\itemize{
\item batch_id - ID of the batch that includes the jobs;
\item jobs_name - a name of jobs on BOINC server;
\item results - computation results (NULL if computation is still incomplete);
The length of this list is equal to the length of the data;
\item jobs_status - human-readable status for each job;
\item status - computation status, may be:
\itemize{
\item "initialization" - jobs have been submitted to the server, but their
status was not requested by update_jobs_status;
\item "in_progress" - BOINC serves jobs;
\item "done" - computations are complete, the results were downloaded;
\item "warning" a recoverable error occurred during the job processing;
\item "error" - a serious error occurred during the job processing;
\item "aborted" - processing was canceled using the
\link[=cancel_jobs]{cancel_jobs} function.
}
}
}
\description{
Update status for jobs and get result for complete jobs.
}
\details{
This function communicates with the boinc server and gets the state for each
job. If the job has already been completed, and its result has been
downloaded and processed, then it is skipped. After the last result is
downloaded, the jobs data is deleted from the server, and status field in the
return value is set to "done".
\subsection{Errors and warnings}{

When errors occur, execution can be stopped with the following messages:
\itemize{
\item for unknown connections:
\itemize{
\item "Unknown protocol."
}
\item for any connection:
\itemize{
\item "All jobs have already been canceled."
\item "All results have already been received."
}
}

This function can output the following warnings:
\itemize{
\item for any connection:
\itemize{
\item Failed to download the result: "\code{<}error message\code{>}"
}
}
}
}
\examples{
\dontrun{

# Function for data processing:
fun = function(val)
{
   ...
}

# Data for processing:
data = list(...)

# Connection to the BOINC server:
con = create_connection(...)

# Send jobs to BOINC server:
jobs = create_jobs(con, fun, data)

# Get status for jobs:
jobs = update_jobs_status(con, jobs)

# Release resources:
close_connection(con)
}
}
